requirements.txt
pandas==1.2.3
requests==2.24.0
omdb==0.10.1
streamlit==0.78.0
Pillow==8.1.2
python-dotenv==0.15.0
scikit_learn==0.24.1
